+++ lib/compiler/gdc/lifetime.d
@@ -838,7 +838,7 @@ extern (C) byte[] _d_arraycatnT(TypeInfo ti, uint n, ...)
         }
     }
 
-    *cast(int *)&a = length; // jam length
+    *cast(size_t *)&a = length; // jam length
     //a.length = length;
     return a;
 }
