+++ lib/compiler/gdc/mars.h
@@ -10,6 +10,7 @@
  */
 
 #include <stddef.h>
+#include <sys/types.h>
 
 #if __cplusplus
 extern "C" {
@@ -28,7 +29,7 @@ typedef struct Interface
 {
     struct ClassInfo *classinfo;
     struct Vtbl vtbl;
-    int offset;
+    ptrdiff_t offset;
 } Interface;
 
 typedef struct Object
